# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/PostRequestDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class PostRequestDTO {

    private Long amountOfCountryToSave;

}